Integrate text boxes with menu selections.
+ Refactor the Rectangle class so that it accepts integers rather than
floats (also to Kotlin). Slick's Graphics class is inconsistent in its
behaviour when drawing shapes and lines that have fractional positional
values. Since it doesn't draw in between pixels, it's better for the
Rectangle class to only accept integers for more predictable behaviour.

+++ java/believe/gui/TextAlignment.kt
@@ -5,18 +5,18 @@ import believe.geometry.Rectangle
 /** Contains enums for aligning text across multiple dimensions. */
 object TextAlignment {
     /** Enums for aligning text on the x axis. */
-    enum class Horizontal(internal val calculateXPosition: (Int, Rectangle) -> Float) {
+    enum class Horizontal(internal val calculateXPosition: (Int, Rectangle) -> Int) {
         /** Aligns text with the left side of its containing component. */
-        LEFT({ _, rect -> rect.x }),
+        LEFT({ _, rect -> rect.x.toInt() }),
         /** Aligns text with the center of its containing component. */
-        CENTERED({ textWidth, rect -> rect.centerX - (textWidth / 2f) });
+        CENTERED({ textWidth, rect -> rect.centerX.toInt() - textWidth / 2 });
     }
 
     /** Enums for aligning text on the y axis. */
-    enum class Vertical(internal val calculateYPosition: (Int, Rectangle) -> Float) {
+    enum class Vertical(internal val calculateYPosition: (Int, Rectangle) -> Int) {
         /** Aligns text with the top of its containing component. */
-        TOP({ _, rect -> rect.y }),
+        TOP({ _, rect -> rect.y.toInt() }),
         /** Aligns text with the middle of its containing component. */
-        MIDDLE({ totalTextFragmentHeight, rect -> rect.centerY - (totalTextFragmentHeight / 2f) });
+        MIDDLE({ totalTextFragmentHeight, rect -> rect.centerY.toInt() - (totalTextFragmentHeight / 2) });
     }
 }
